Move ArrayOfVectorViews into submodule CollectionsOfVectorViews
